requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.4.26
charset-normalizer==3.4.1
Django==4.2.20
django-bootstrap-v5==1.0.11
django-filter==25.1
django-push-notifications==3.2.1
django-rest-passwordreset==1.5.0
djangorestframework==3.16.0
idna==3.10
pillow==11.1.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
